requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
plotly
xgboost
imbalanced-learn
streamlit
pylint
pytest
black
isort
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
pickle
python-multipart